updated Sample app
