--- Sample/AndroidManifest.xml
@@ -1,61 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.phunware.core.sample"
-    android:versionCode="1"
-    android:versionName="1.0" >
-
-    <uses-sdk
-        android:minSdkVersion="8"
-        android:targetSdkVersion="17" />
-    
-    <!-- Necessary for core to communicate with MaaS server -->
-	<uses-permission android:name="android.permission.INTERNET" />
-	<!-- Used to efficiently communicate with MaaS Servers -->
-	<uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
-	
-	<!-- Used to get location analytic data -->
-    <uses-permission android:name="android.permission.ACCESS_FINE_LOCATION" />
-    <uses-permission android:name="android.permission.ACCESS_COARSE_LOCATION" />
-
-    <application
-        android:name="com.phunware.core.sample.MyApplication"
-        android:allowBackup="true"
-        android:icon="@drawable/ic_launcher"
-        android:label="@string/app_name"
-        android:theme="@style/AppTheme" >
-        
-        <!-- Necessary for core to communicate with MaaS server -->
-	    <service android:name="com.phunware.core.internal.CoreService" />
-	    <!-- Necessary to generate a UDID -->
-	    <service android:name="com.OpenUDID.OpenUDID_service">
-			 <intent-filter>
-				<action android:name="com.openudid.GETUDID" />
-			</intent-filter>
-		</service>
-		<!-- Used by Core for efficient Analytic Caching -->
-		<receiver android:name="com.phunware.core.internal.CoreReceiver">
-		    <intent-filter>
-		        <action android:name="android.net.conn.CONNECTIVITY_CHANGE" />
-		    </intent-filter>
-		</receiver>
-        
-        <activity
-            android:name="com.phunware.core.sample.CoreDemo"
-            android:label="@string/app_name" >
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
-        </activity>
-        <activity
-            android:name="com.phunware.core.sample.SecondActivity"
-            android:label="@string/title_activity_second"
-            android:parentActivityName="com.phunware.core.sample.CoreDemo" >
-            <meta-data
-                android:name="android.support.PARENT_ACTIVITY"
-                android:value="com.phunware.core.sample.MainActivity" />
-        </activity>
-    </application>
-
-</manifest>
