+++ Sample/README.md
@@ -6,6 +6,7 @@ This sample demonstrates proper usage of the MaaS Core SDK
 
 ## Setup
 
-1. Import into an IDE as an existing Android Project. This will generate the necessary local files and directories (gen, bin, .project, .classpath, etc).
-2. Copy the MaaS Core SDK into the `libs` directory.
-3. Run as Android Project
+1. Import into an IDE as an existing Android Project.
+2. Modify credentials for top level build.gradle file. Enter a `username` and a `password`.
+3. In `strings.xml`, modify `app_appid`, `app_accesskey`, and `app_signaturekey` with the values that were configured for your app in the MaaS Portal.
+4. Run as Android Project
