+++ Sample/app/src/main/AndroidManifest.xml
@@ -1,0 +1,31 @@
+<?xml version="1.0" encoding="utf-8"?>
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="com.phunware.core.sample" >
+
+    <application
+        android:name="com.phunware.core.sample.MyApplication"
+        android:allowBackup="true"
+        android:icon="@mipmap/ic_launcher"
+        android:label="@string/app_name"
+        android:supportsRtl="true"
+        android:theme="@style/AppTheme" >
+        <activity
+            android:name="com.phunware.core.sample.CoreDemo"
+            android:label="@string/app_name" >
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
+        </activity>
+        <activity
+            android:name="com.phunware.core.sample.SecondActivity"
+            android:label="@string/title_activity_second"
+            android:parentActivityName="com.phunware.core.sample.CoreDemo" >
+            <meta-data
+                android:name="android.support.PARENT_ACTIVITY"
+                android:value="com.phunware.core.sample.CoreDemo" />
+        </activity>
+    </application>
+
+</manifest>
