+++ Sample/app/src/main/java/com/phunware/core/sample/SecondActivity.java
@@ -1,0 +1,54 @@
+package com.phunware.core.sample;
+
+import android.annotation.TargetApi;
+import android.app.Activity;
+import android.os.Build;
+import android.os.Bundle;
+import android.support.v4.app.NavUtils;
+import android.view.MenuItem;
+
+import com.phunware.core.PwCoreSession;
+
+public class SecondActivity extends Activity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_second);
+    }
+
+    @Override
+    protected void onStart() {
+        super.onStart();
+		/*
+		 * Start Core Session
+		 */
+        PwCoreSession.getInstance().activityStartSession(this);
+    }
+
+    @Override
+    protected void onStop() {
+        super.onStop();
+		/*
+		 * Stop Core Session
+		 */
+        PwCoreSession.getInstance().activityStopSession(this);
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        switch (item.getItemId()) {
+            case android.R.id.home:
+                // This ID represents the Home or Up button. In the case of this
+                // activity, the Up button is shown. Use NavUtils to allow users
+                // to navigate up one level in the application structure. For
+                // more details, see the Navigation pattern on Android Design:
+                //
+                // http://developer.android.com/design/patterns/navigation.html#up-vs-back
+                //
+                NavUtils.navigateUpFromSameTask(this);
+                return true;
+        }
+        return super.onOptionsItemSelected(item);
+    }
+}
