+++ Sample/app/src/main/res/layout/activity_second.xml
