--- Sample/res/values/strings.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-
-    <string name="app_name">PwCore Demo</string>
-    <string name="app_appid">appid</string>
-    <string name="app_accesskey">abcd12345</string>
-    <string name="app_signaturekey">098765fghjkl</string>
-    <string name="app_encryptionkey">65as6d54f065sd4f0</string>
-    <string name="title_activity_second">SecondActivity</string>
-
-</resources>
