+++ Sample/settings.gradle
@@ -1,0 +1,1 @@
+include ':app'
